feat: Add routes for notifications and individual posts
Introduces new routes for `/notifications` and `/post/:id` to allow users to view their notifications and specific posts respectively.

Also includes updates to the header component to include a notification icon and functionality, and adjusts the color theme for dark mode in `index.html` for a more consistent aesthetic. The theme color for dark mode has been updated from `#303030` to `#121212`.

Finally, enhances the `PostCard` component with a share functionality, allowing users to share posts via the Web Share API or by copying the link to their clipboard. Notification fetching has also been adjusted to fetch the last 50 notifications instead of 30.